Feat: base do modulo de envio
