--- ts/cel/email-cel.ts
@@ -1,3 +1,0 @@
-interface email {
-
-}
