--- ts/franco/franco.interface.ts
@@ -1,6 +1,0 @@
-interface EmailParams {
-    nome: string;
-    cursos: string[],
-}
-
-export default EmailParams
