+++ ts/src/cel/email-cel.ts
@@ -1,0 +1,3 @@
+interface email {
+
+}
