+++ ts/src/franco/franco.interface.ts
@@ -1,0 +1,6 @@
+interface EmailParams {
+    nome: string;
+    cursos: string[],
+}
+
+export default EmailParams
